requirements.txt
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
PyJWT==1.7.1
SQLAlchemy==1.3.19
Werkzeug==1.0.1
